+++ src/Symfony/Component/Yaml/Command/LintCommand.php
@@ -58,7 +58,9 @@ class LintCommand extends Command
             ->setDescription(self::$defaultDescription)
             ->addArgument('filename', InputArgument::IS_ARRAY, 'A file, a directory or "-" for reading from STDIN')
             ->addOption('format', null, InputOption::VALUE_REQUIRED, 'The output format')
-            ->addOption('parse-tags', null, InputOption::VALUE_NONE, 'Parse custom tags')
+            ->addOption('exclude', 'e', InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Path(s) to exclude')
+            ->addOption('config', 'c', InputOption::VALUE_REQUIRED, 'Path to YAML lint config')
+            ->addOption('parse-tags', null, InputOption::VALUE_NEGATABLE, 'Parse custom tags', null)
             ->setHelp(<<<EOF
 The <info>%command.name%</info> command lints a YAML file and outputs to STDOUT
 the first encountered syntax error.
@@ -66,6 +68,10 @@ the first encountered syntax error.
 You can validates YAML contents passed from STDIN:
 
   <info>cat filename | php %command.full_name% -</info>
+
+You can validate YAML contents passed from config file:
+
+  <info>php %command.full_name% --config=.yamllint.yml</info>
 
 You can also validate the syntax of a file:
 
@@ -76,6 +82,10 @@ Or of a whole directory:
   <info>php %command.full_name% dirname</info>
   <info>php %command.full_name% dirname --format=json</info>
 
+You can also exclude one or more specific files:
+
+  <info>php %command.full_name% dirname --exclude="dirname/foo.yml" --exclude="dirname/bar.yml"</info>
+
 EOF
             )
         ;
@@ -85,7 +95,24 @@ EOF
     {
         $io = new SymfonyStyle($input, $output);
         $filenames = (array) $input->getArgument('filename');
+        $config = $input->getOption('config');
+        $excludes = $input->getOption('exclude');
         $this->format = $input->getOption('format');
+        $flags = $input->getOption('parse-tags');
+
+        if ($config) {
+            $lintConfig = $this->parseLintConfig($config);
+
+            $filenames = array_merge($lintConfig['includes'], $filenames);
+            $excludes = array_merge($lintConfig['excludes'], $excludes);
+
+            if (null === $this->format && isset($lintConfig['format'])) {
+                $this->format = $lintConfig['format'];
+            }
+            if (null !== $flags && isset($lintConfig['parse-tags'])) {
+                $flags = $lintConfig['parse-tags'];
+            }
+        }
 
         if ('github' === $this->format && !class_exists(GithubActionReporter::class)) {
             throw new \InvalidArgumentException('The "github" format is only available since "symfony/console" >= 5.3.');
@@ -96,8 +123,9 @@ EOF
             $this->format = class_exists(GithubActionReporter::class) && GithubActionReporter::isGithubActionEnvironment() ? 'github' : 'txt';
         }
 
+        $flags = $flags ? Yaml::PARSE_CUSTOM_TAGS : 0;
+
         $this->displayCorrectFiles = $output->isVerbose();
-        $flags = $input->getOption('parse-tags') ? Yaml::PARSE_CUSTOM_TAGS : 0;
 
         if (['-'] === $filenames) {
             return $this->display($io, [$this->validate(file_get_contents('php://stdin'), $flags)]);
@@ -114,7 +142,9 @@ EOF
             }
 
             foreach ($this->getFiles($filename) as $file) {
-                $filesInfo[] = $this->validate(file_get_contents($file), $flags, $file);
+                if (!\in_array($file->getPathname(), $excludes, true)) {
+                    $filesInfo[] = $this->validate(file_get_contents($file), $flags, $file);
+                }
             }
         }
 
@@ -266,5 +296,25 @@ EOF
         }
 
         return $default($fileOrDirectory);
+    }
+
+    private function parseLintConfig(string $config): array
+    {
+        if (!file_exists($config)) {
+            throw new RuntimeException(sprintf('YAML Lint config "%s" not found.', $config));
+        }
+
+        $result = (new Parser())
+            ->parseFile($config);
+
+        $allowedProperties = ['format', 'includes', 'excludes', 'parse-tags'];
+        if (!\is_array($result) || !empty(array_diff(array_keys($result), $allowedProperties))) {
+            throw new RuntimeException(sprintf('Invalid YAML lint config "%s".', $config));
+        }
+
+        $result['includes'] = $result['includes'] ?? [];
+        $result['excludes'] = $result['excludes'] ?? [];
+
+        return $result;
     }
 }